requirements_gui.txt
PyQt5>=5.15.0
PyYAML>=6.0

